netdog: refactor cli code into modules
Since `main.rs` was starting to get large and unwieldy, and additional
code is expected to be added, this commit refactors all the CLI code
into separate modules grouped under a `cli` module.
